refactor(plugins): move Discord plugin to src/plugins/discord/ for unified compilation
Move the Discord plugin from plugins/discord/ into src/plugins/discord/ so it
compiles alongside the main package. This ensures the plugin works both from
source (pnpm dev) and npm install (dist/plugins/discord/) without fragile
relative imports or separate build steps.

- Move all plugin files to src/plugins/discord/
- Update imports from ../../src/core/ to ../../core/ (correct relative depth)
- Update plugin discovery path in runtime.ts
- Copy isotopes.plugin.json to dist/ during build
- Fix .gitignore to only ignore top-level /plugins/ (user-installed)
