+++ src/plugins/discord/config.ts
