+++ src/plugins/discord/isotopes.plugin.json
