+++ src/plugins/discord/message-metadata.ts
